// src/app/public/modules/date-range-picker/types/date-range-format-type.ts
export enum SkyDateRangeFormatType {
  SpecificRange = 0,
  Before = 1,
  After = 2,
  AtAnyTime = 3,
  Yesterday = 4,
  Today = 5,
  Tomorrow = 6,
  LastWeek = 7,
  ThisWeek = 8,
  NextWeek = 9,
  LastMonth = 10,
  ThisMonth = 11,
  NextMonth = 12,
  LastQuarter = 13,
  ThisQuarter = 14,
  NextQuarter = 15,
  LastYear = 16,
  ThisYear = 17,
  NextYear = 18,
  LastFiscalYear = 19,
  ThisFiscalYear = 20,
  NextFiscalYear = 21
}